requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.1
Flask-MySQL==1.5.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
PyMySQL==1.0.3
pyodbc==4.0.34
Werkzeug==2.3.2
zipp==3.15.0
